Update code comments
