+++ src/video/binktypes.ts
@@ -1,3 +1,15 @@
+/**
+ * BinkTypes class.
+ * 
+ * KotOR JS - A remake of the Odyssey Game Engine that powered KotOR I & II
+ * 
+ * @file BinkTypes.ts
+ * @see https://github.com/FFmpeg/FFmpeg/blob/release/3.4/libavcodec/bink.h
+ * @autthor Lachjames <https://github.com/Lachjames> (Ported from FFmpeg)
+ * @author KobaltBlu <https://github.com/KobaltBlu> (Modified for KotOR JS)
+ * @license {@link https://www.gnu.org/licenses/gpl-3.0.txt|GPLv3}
+ */
+
 export type FourCC = string; // 4-char tag like 'BIKb' or 'KB2a'
 
 export interface Rational {
